# docs/autogenerated/openapi.yaml
openapi: 3.0.1
info:
  title: REST API
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: https://localhost:8080
  description: Generated server url
...
paths:
...
  /repair-management/v2/schedules:
    get:
      tags:
      - Repair-Management
      summary: Get schedules
      description: Get schedules
      operationId: get-schedules
      parameters:
      - name: keyspace
        in: query
        description: "Filter schedules based on this keyspace, mandatory if 'table'\
          \ is provided."
        required: false
        schema:
          type: string
      - name: table
        in: query
        description: Filter schedules based on this table.
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
  /repair-management/v2/schedules/{id}:
    get:
      tags:
      - Repair-Management
      summary: Get schedules matching the id.
      description: Get schedules matching the id.
      operationId: get-schedules-by-id
      parameters:
      - name: id
        in: path
        description: The id of the schedule.
        required: true
        schema:
          type: string
      - name: full
        in: query
        description: Decides if a 'full schedule' should be returned.
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
...
  /repair-management/v2/repairInfo:
    get:
      tags:
      - Repair-Management
      summary: Get repair information
      description: "Get repair information, if keyspace and table are provided while\
        \ duration and since are not, the duration will default to GC_GRACE_SECONDS\
        \ of the table. This operation might take time depending on the provided params\
        \ since it's based on the repair history."
      operationId: get-repair-info
      parameters:
      - name: keyspace
        in: query
        description: "Only return repair-info matching the keyspace, mandatory if\
          \ 'table' is provided."
        required: false
        schema:
          type: string
      - name: table
        in: query
        description: Only return repair-info matching the table.
        required: false
        schema:
          type: string
      - name: since
        in: query
        description: "Since time, can be specified as ISO8601 date or as milliseconds\
          \ since epoch. Required if keyspace and table or duration is not specified."
        required: false
        schema:
          type: string
      - name: duration
        in: query
        description: "Duration, can be specified as either a simple duration like\
          \ '30s' or as ISO8601 duration 'pt30s'. Required if keyspace and table or\
          \ since is not specified."
        required: false
        schema:
          type: string
      - name: isLocal
        in: query
        description: Decides if the repair-info should be calculated for the local
          node only.
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairInfo'
...
components:
  schemas:
...
    Schedule:
      required:
      - config
      - id
      - keyspace
      - lastRepairedAtInMs
      - nextRepairInMs
      - repairedRatio
      - status
      - table
      type: object
      properties:
        id:
          type: string
          format: uuid
        keyspace:
          type: string
        table:
          type: string
        status:
          type: string
          enum:
          - COMPLETED
          - ON_TIME
          - LATE
          - OVERDUE
          - BLOCKED
        repairedRatio:
          maximum: 1
          minimum: 0
          type: number
          format: double
        lastRepairedAtInMs:
          type: integer
          format: int64
        nextRepairInMs:
          type: integer
          format: int64
        config:
          $ref: '#/components/schemas/ScheduleConfig'
        virtualNodeStates:
          type: array
          items:
            $ref: '#/components/schemas/VirtualNodeState'
    ScheduleConfig:
      required:
      - errorTimeInMs
      - intervalInMs
      - parallelism
      - unwindRatio
      - warningTimeInMs
      type: object
      properties:
        intervalInMs:
          minimum: 0
          type: integer
          format: int64
        unwindRatio:
          minimum: 0
          type: number
          format: double
        warningTimeInMs:
          minimum: 0
          type: integer
          format: int64
        errorTimeInMs:
          minimum: 0
          type: integer
          format: int64
        parallelism:
          type: string
          enum:
          - PARALLEL
    VirtualNodeState:
      required:
      - endToken
      - lastRepairedAtInMs
      - repaired
      - replicas
      - startToken
      type: object
      properties:
        startToken:
          minimum: -9223372036854775808
          type: integer
          format: int64
        endToken:
          maximum: 9223372036854775807
          type: integer
          format: int64
        replicas:
          uniqueItems: true
          type: array
          items:
            type: string
        lastRepairedAtInMs:
          minimum: 0
          type: integer
          format: int64
        repaired:
          type: boolean
    RepairInfo:
      required:
      - repairStats
      - sinceInMs
      - toInMs
      type: object
      properties:
        sinceInMs:
          minimum: 0
          type: integer
          format: int64
        toInMs:
          minimum: 0
          type: integer
          format: int64
        repairStats:
          type: array
          items:
            $ref: '#/components/schemas/RepairStats'
    RepairStats:
      required:
      - keyspace
      - repairTimeTakenMs
      - repairedRatio
      - table
      type: object
      properties:
        keyspace:
          type: string
        table:
          type: string
        repairedRatio:
          maximum: 1
          minimum: 0
          type: number
          format: double
        repairTimeTakenMs:
          minimum: 0
          type: integer
          format: int64
...